requirements.txt
streamlit==1.39.0
langchain==0.3.7
langchain-community==0.3.5
openai==1.53.0
langchain-openai==0.2.5
streamlit-authenticator==0.4.1
redis==5.2.0
pypdf==5.1.0
docarray==0.40.0
unstructured==0.16.6
python-docx==1.1.2
python-dotenv==1.0.1